helix-core/surround: `Syntax` aware
Prior to this change, attempting to replace the double quotes in
`" asdf "` would fail with "Cursor on ambiguous surround pair".

Now, after running e.g. `:set-language rust` in a file with the contents
`" asdf "`, it is now possible to replace the double quotes with `mr` as
you could with any other pair.

This change essentially makes any pair that you can use `mm` on to go
to the matching pair also able to be replaced with `mr` or deleted with
`md` (whether they be built-in, or provided by a tree-sitter grammar).